% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_culture_level}
\alias{data_culture_level}
\title{DATASET_TITLE}
\format{
A data frame with 60 rows and 25 variables:
\describe{
\item{\code{owc_name}}{character COLUMN_DESCRIPTION}
\item{\code{owc_id}}{character COLUMN_DESCRIPTION}
\item{\code{sccs_id}}{character COLUMN_DESCRIPTION}
\item{\code{Documents}}{double COLUMN_DESCRIPTION}
\item{\code{Region}}{character COLUMN_DESCRIPTION}
\item{\code{Subregion}}{character COLUMN_DESCRIPTION}
\item{\code{Subsistence Type}}{character COLUMN_DESCRIPTION}
\item{\code{Writing_SCCSv149}}{character COLUMN_DESCRIPTION}
\item{\code{Fixity_residence_SCCSv150}}{character COLUMN_DESCRIPTION}
\item{\code{Agricultre_SCCSv151}}{character COLUMN_DESCRIPTION}
\item{\code{Urbanization_SCCSv152}}{character COLUMN_DESCRIPTION}
\item{\code{Labor_specialization_SCCSv153}}{character COLUMN_DESCRIPTION}
\item{\code{Land_transport_SCCSv1554}}{character COLUMN_DESCRIPTION}
\item{\code{Money_SCCSv155}}{character COLUMN_DESCRIPTION}
\item{\code{Density_population_SCCSv156}}{character COLUMN_DESCRIPTION}
\item{\code{Political_integration_SCCSv157}}{character COLUMN_DESCRIPTION}
\item{\code{Social_stratification_SCCSv158}}{character COLUMN_DESCRIPTION}
\item{\code{Hunting_SCCSv204}}{integer COLUMN_DESCRIPTION}
\item{\code{Food_storage_SCCSv20}}{character COLUMN_DESCRIPTION}
\item{\code{community_organization_recode}}{character COLUMN_DESCRIPTION}
\item{\code{domestic_organization_recode}}{character COLUMN_DESCRIPTION}
\item{\code{marital_residence_recode}}{character COLUMN_DESCRIPTION}
\item{\code{polygamy_recode}}{character COLUMN_DESCRIPTION}
\item{\code{cousin_marriage_recode}}{character COLUMN_DESCRIPTION}
\item{\code{lineage_organization_recode}}{character COLUMN_DESCRIPTION}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
data_culture_level
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
