% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_paragraph}
\alias{data_paragraph}
\title{DATASET_TITLE}
\format{
A data frame with 8775 rows and 8 variables:
\describe{
\item{\code{uuid}}{character COLUMN_DESCRIPTION}
\item{\code{coder}}{character COLUMN_DESCRIPTION}
\item{\code{offense_present}}{character COLUMN_DESCRIPTION}
\item{\code{unspecified}}{character COLUMN_DESCRIPTION}
\item{\code{secondparty}}{character COLUMN_DESCRIPTION}
\item{\code{thirdparty}}{character COLUMN_DESCRIPTION}
\item{\code{thirdparty_kin}}{character COLUMN_DESCRIPTION}
\item{\code{thirdparty_partner}}{character COLUMN_DESCRIPTION}
\item{\code{thirdparty_general}}{character COLUMN_DESCRIPTION}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
data_paragraph
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
